Dynamically fetch rules on page load
