+++ src/components/Rule.js
@@ -4,24 +4,7 @@ import { useLocation } from 'react-router-dom';
 import { getData } from '../apiCalls'
 import Page from './Page'
 
-const Rule = ({ rule }) => {
-  const location = useLocation();
-
-  const [page, setPage] = useState({});
-
-  const paths = {
-    'gameplay-basics': 'rule-sections',
-    'abilities': 'ability-scores',
-    'skills': 'rules',
-    'races': 'races',
-    'classes': 'classes'
-  }
-
-  useEffect(() => {
-    getData(paths[location.pathname.split('/')[1]], location.pathname.split('/')[2])
-    .then(res => setPage(res))
-  }, [])
-
+const Rule = ({page}) => {
 
   const formatSubheader = (phrase, index) => {
     const words = phrase.split(' ');
@@ -30,16 +13,13 @@ const Rule = ({ rule }) => {
   }
 
   const formatContent = () => {
-    console.log('page:', page)
     return page.desc.split('\n').filter(desc => desc).map((desc, index) => desc.includes('#') ? formatSubheader(desc, index) : <p key={index}>{desc}</p>)
   }
 
   return (
-    <main>
-      <h2 className='main__heading'>{page.name}</h2>
-      <div className='main__divider'></div>
+    <>
       {formatContent()}
-    </main>
+    </>
   )
 }
